Restore original peer id prefix.
There is no precedent for forcing a development branch to have
a different peer id, either for any other github forks or personal
svn repositories used by users and contributors. The "Z" in the
peer id already signifies that the build is unsupported by the
development team and therefore may contain unsound or experimental
changes.